Refactor: logic lasses not dependent on GUI
ProfileLogic methods made static
